Control/UAV/Ardupilot: Made "Seconds before Waypoint" float
To enable more fine tuning

+++ src/Control/UAV/Ardupilot/Task.cpp
@@ -137,7 +137,7 @@ namespace Control
         //! Has Power Module
         bool pwrm;
         //! WP seconds before reach
-        int secs;
+        float secs;
         //! WP Copter: Minimum wp switch radius
         float cp_wp_radius;
         //! RC setup
